+++ pages/_app.js
@@ -3,7 +3,6 @@ import React, { useState } from 'react';
 import Head from 'next/head';
 import PropTypes from 'prop-types';
 import withReduxSaga from 'next-redux-saga';
-import 'antd/dist/antd.css';
 import LinearProgress from '@material-ui/core/LinearProgress';
 
 
@@ -18,6 +17,7 @@ const WebT = ({ Component }) => {
       <Head>
         <title>Hello</title>
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
+        <script src="/node_modules/quill-image-resize-module/image-resize.min.js"></script>
 
       </Head>
       <Component />
