+++ sagas/post.js
@@ -7,7 +7,7 @@ import {
   ADD_POST_REQUEST,
   ADD_POST_SUCCESS,
 } from '../reducers/post';
-import { LOAD_POSTS_REQUEST } from '../reducers/posts';
+import { LOAD_POSTS_FAILURE, LOAD_POSTS_REQUEST, LOAD_POSTS_SUCCESS } from '../reducers/posts';
 
 function loadPostAPI(blogId) {
   return axios.post('http://localhost:3065/api/post/',{blogId:blogId});
